Update prediction_app.R
Added an input_panel branch. Not sure if we care enough to have branches or just want to suggest/commit changes to the master branch. 

For this, I changed default values from TRUE to NULL, so that the input panel is fully blank unless the clinician inputs checks boxes or inputs any information. I also messed with the formatting somewhat to add our names and a title, and have the sidebar panel in a separate panel. I've included some placeholder text for the results in the main panel.  I also named the input check boxes, so that I can work with the specific inputs in calculating predicted probabilities.