Update prediction_app.R
Added an additional `datatable` object with some of the initial output. The table is editable. I still need to add the "Classification" column for "Seizure/No seizure" prediction status. I also still need to figure out the best way to update the sensitivity and specificity values based on user-selected threshold