Update prediction_app.R
Few updates. Some cosmetic. I added LaTeX coding of the logistic regression formula  and predicted probability calculation. Not sure if we want to keep these, but have added them for now. I still need to update the header instructions/introduction, but will work on that later tonight in a very small commit. Added the slider for threshold, moving sensitivity/specificity based on the selected threshold, and quantile columns within all ~4,000 patients, plus those few thousand without PTS events and those few hundred with. 

Just a note, the threshold value is the object `input$thresholdslider`, if you use that at all in the plots!