Update prediction_app.R
I copy and pasted my entire script, so there are not as many substantive edits ad it looks like. The biggest changes I made included:
     I changed the plotly object to simply plot the density of predicted probabilities of PTE vs non PTE, and added a v-line for the predicted probability as well as the threshold
     Editing the quantile_fun function to allow for value inputs, which allowed me to calculate a reactive input geom_vline for the plotly object with percentiles
     Editing some of the LaTeX wrapping to remove the HTML coding around the text adn simply pass LaTeX syntax to a uiOutput object

I'm still trying to edit the plotly object to allow for custom hoverover text. I've attempted to use tooltip and hoverover options, but they don't react correctly for some reason, even following a number of StackOverflow posts and the help documentation. Nothing we need to do by tomorrow, it's not that important for the app and teh relevant information is already contained in the table output